Move classes to dedicated files per Kotlin best practices
