+++ src/main/kotlin/tech/relaycorp/relaynet/RelaynetException.kt
@@ -1,0 +1,3 @@
+package tech.relaycorp.relaynet
+
+abstract class RelaynetException(message: String) : Exception(message)
