+++ src/main/kotlin/tech/relaycorp/relaynet/wrappers/KeyException.kt
@@ -1,0 +1,5 @@
+package tech.relaycorp.relaynet.wrappers
+
+import tech.relaycorp.relaynet.RelaynetException
+
+class KeyException(message: String) : RelaynetException(message)
