+++ src/main/kotlin/tech/relaycorp/relaynet/wrappers/x509/CertificateException.kt
@@ -1,0 +1,5 @@
+package tech.relaycorp.relaynet.wrappers.x509
+
+import tech.relaycorp.relaynet.RelaynetException
+
+class CertificateException(message: String) : RelaynetException(message)
